move Interval from util to utils
