+++ api/src/main/java/de/oliver/fancynpcs/api/Npc.java
@@ -6,8 +6,8 @@ import de.oliver.fancylib.RandomUtils;
 import de.oliver.fancylib.translations.Translator;
 import de.oliver.fancynpcs.api.events.NpcInteractEvent;
 import de.oliver.fancynpcs.api.events.NpcInteractEvent.InteractionType;
-import de.oliver.fancynpcs.api.util.Interval;
-import de.oliver.fancynpcs.api.util.Interval.Unit;
+import de.oliver.fancynpcs.api.utils.Interval;
+import de.oliver.fancynpcs.api.utils.Interval.Unit;
 import me.dave.chatcolorhandler.ChatColorHandler;
 import me.dave.chatcolorhandler.ModernChatColorHandler;
 import me.dave.chatcolorhandler.parsers.custom.PlaceholderAPIParser;
