+++ src/main/java/de/oliver/fancynpcs/commands/npc/InfoCMD.java
@@ -5,8 +5,8 @@ import de.oliver.fancylib.translations.message.Message;
 import de.oliver.fancylib.translations.message.SimpleMessage;
 import de.oliver.fancynpcs.FancyNpcs;
 import de.oliver.fancynpcs.api.Npc;
-import de.oliver.fancynpcs.api.util.Interval;
-import de.oliver.fancynpcs.api.util.Interval.Unit;
+import de.oliver.fancynpcs.api.utils.Interval;
+import de.oliver.fancynpcs.api.utils.Interval.Unit;
 import de.oliver.fancynpcs.utils.GlowingColor;
 import org.bukkit.Location;
 import org.bukkit.command.CommandSender;
