+++ src/main/java/de/oliver/fancynpcs/commands/npc/InteractionCooldownCMD.java
@@ -4,8 +4,8 @@ import de.oliver.fancylib.translations.Translator;
 import de.oliver.fancynpcs.FancyNpcs;
 import de.oliver.fancynpcs.api.Npc;
 import de.oliver.fancynpcs.api.events.NpcModifyEvent;
-import de.oliver.fancynpcs.api.util.Interval;
-import de.oliver.fancynpcs.api.util.Interval.Unit;
+import de.oliver.fancynpcs.api.utils.Interval;
+import de.oliver.fancynpcs.api.utils.Interval.Unit;
 import de.oliver.fancynpcs.commands.exceptions.ReplyingParseException;
 import it.unimi.dsi.fastutil.Pair;
 import org.bukkit.command.CommandSender;
